+++ src/request/request-factory.js
@@ -8,10 +8,15 @@ class RequestFactory {
         internal(this).endpoint = `${url}/api/v${version}/content/${channel}`;
         internal(this).locale = null;
         internal(this).defaultResources = [];
+        internal(this).origin = null;
     }
 
     set locale(locale) {
         internal(this).locale = locale;
+    }
+
+    set origin(origin) {
+        internal(this).origin = origin;
     }
 
     addDefaultResource(name, value) {
@@ -20,8 +25,16 @@ class RequestFactory {
 
     create() {
         const _internal = internal(this);
+        const headers = [];
 
-        return new Request(_internal.method, _internal.endpoint, _internal.locale, _internal.defaultResources);
+        if ('string' === typeof _internal.origin && '' !== _internal.origin) {
+            headers.push({
+                name: 'X-Amp-Origin',
+                value: _internal.origin,
+            });
+        }
+
+        return new Request(_internal.method, _internal.endpoint, _internal.locale, _internal.defaultResources, headers);
     }
 }